+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
+## [Unreleased] - YYYY-MM-DD
+
+### Added
+
+- Authentication support with bearer tokens and login endpoint
+- Fix docker and build issues
+
 ## [0.1.0] - 2025-10-14
 
 ### Added